Fixed counter positioning

+++ src/game/counter.h
@@ -20,7 +20,7 @@ struct counter {
 typedef struct counter *counter_t;
 
 counter_t counter_create(void);
-void counter_update_x(counter_t counter, uint x);
+void counter_update_x(counter_t counter, float x);
 void counter_render(const counter_t counter, mat4 projection);
 void counter_free(counter_t counter);
